requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
datasets==2.14.0
pycocotools==2.0.6
numpy==1.24.3
opencv-python==4.8.0.74
pyyaml==6.0
Pillow==10.0.0
timm==0.9.12  # Add this line